+++ ledger/ledger-api-test-tool/src/main/scala/com/daml/ledger/api/testtool/LedgerApiTestTool.scala
@@ -15,7 +15,6 @@ import io.grpc.Channel
 import io.grpc.netty.{NegotiationType, NettyChannelBuilder}
 import org.slf4j.LoggerFactory
 
-import scala.collection.compat._
 import scala.concurrent.duration.DurationInt
 import scala.concurrent.{ExecutionContext, Future}
 import scala.util.{Failure, Success}
@@ -92,7 +91,11 @@ object LedgerApiTestTool {
       timeoutScaleFactor = config.timeoutScaleFactor,
       ledgerClockGranularity = config.ledgerClockGranularity,
     )
-    val visibleTests: Vector[LedgerTestSuite] = defaultTests ++ Tests.optional
+    val optionalTests: Vector[LedgerTestSuite] = Tests.optional(
+      timeoutScaleFactor = config.timeoutScaleFactor,
+      ledgerClockGranularity = config.ledgerClockGranularity,
+    )
+    val visibleTests: Vector[LedgerTestSuite] = defaultTests ++ optionalTests
     val allTests: Vector[LedgerTestSuite] = visibleTests ++ Tests.retired
     val allTestCaseNames: Set[String] = allTests.flatMap(_.tests).map(_.name).toSet
     val missingTests = (config.included ++ config.excluded).filterNot(prefix =>
@@ -146,7 +149,7 @@ object LedgerApiTestTool {
       })
 
     val defaultCases = defaultTests.flatMap(_.tests)
-    val allCases = defaultCases ++ Tests.optional.flatMap(_.tests) ++ Tests.retired.flatMap(_.tests)
+    val allCases = defaultCases ++ optionalTests.flatMap(_.tests) ++ Tests.retired.flatMap(_.tests)
 
     val includedTests =
       if (config.included.isEmpty) defaultCases
